requirements/base.txt
Django==1.11.15
django-extensions==2.1.2
requests==2.19.1
django-model-utils==3.1.2
django-cas-client==1.4.0
PyMySQL==0.9.2
django-hijack==2.1.10
django-hijack-admin==2.1.10

# api tools
djangorestframework==3.8.2
Markdown==2.6.11
django-filter==2.0.0
--extra-index-url https://pypi.lib.byu.edu
django-prysm-zombies==3.1
python-bayou==1.5.0
python-configuro==0.4.0
